+++ fsm/types/key.go
@@ -10,7 +10,6 @@ import (
 
 var ReservedIDs = []uint64{
 	lib.UnknownCommitteeId,
-	lib.CanopyCommitteeId,
 	lib.DAOPoolID,
 }